# Trace load and json parsing optimizations