--- a/hta/analyzers/straggler.py
+++ b/hta/analyzers/straggler.py
@@ -9,7 +9,8 @@ import pandas as pd
 import plotly.express as px
 
 from hta.analyzers.timeline import _get_unique_values, plot_timeline_gpu_kernels
-from hta.common.trace import Trace, TraceSymbolTable
+from hta.common.trace import Trace
+from hta.common.trace_symbol_table import TraceSymbolTable
 
 
 def extract_iteration_info(trace: Trace) -> pd.DataFrame: