## What does this PR do?

Trace loading is one of the slowest steps in analysis and consumes the highest memory with 25% traces having >2.7 millions events.
* We propose a solution to speed up trace load time of large trace(s) by 2.6x and more importantly reduce the peak memory usage by 8.5x.

**Why we can do better?**
* Overall trace loading for a large trace is taking 6 GBs of memory. But the json object only takes about 800 MB.
* For large traces a significant fraction of events are “python_function” related and are not used in trace analysis.
Following are results on a large trace
<img width="658" alt="Screenshot 2024-03-26 at 10 49 46 AM" src="https://github.com/facebookresearch/HolisticTraceAnalysis/assets/6922212/8efe7ffa-e75c-4a2f-8b9a-a4baa3517857">

### Proposed solution
* We leverage ijson to filter "python_function" events that are not used in analysis.
* Add on optimizations that use batching, and extracting the args on the fly. This removes most of the memory overhead

Caveat: ijson can use an optimized C based json parser (yajl), if available. The user should have this else python based parsing is much slower that simply using

### Code Details
1. Moved the symbol table and json parsing (trace_parser.py) into two different files.
2. Proposed a `parser_backend` option in `ParserCfg`.
3. Added ijson based parsing in trace_parser.py (main set of changes are here.
4. Added simple unit test for ijson based parsing.
5. Auto-detect best backend (this is not used right now to keep json backend as default)

## Testing

### Unit test
* `pytest tests/test_symbol_table.py`
* `pytest tests/test_trace_parse.py -k TraceParseIjson`
* `pytest tests` 

Checking automatic detection of backend.

### Benchmarking results

Run
`python3 benchmarks/trace_load_benchmark.py -v -p 1 -l 1`
```
parse[json:tests/data/vision_transformer/rank-1.json.gz]: Mean +- std dev: 6.53 sec +- 0.36 sec
parse[ijson:tests/data/vision_transformer/rank-1.json.gz]: Mean +- std dev: 7.57 sec +- 0.02 sec
parse[ijson_batched:tests/data/vision_transformer/rank-1.json.gz]: Mean +- std dev: 7.89 sec +- 0.24 sec
parse[ijson_batch_and_compress:tests/data/vision_transformer/rank-1.json.gz]: Mean +- std dev: 5.70 sec +- 0.21 sec
parse[json:../../load_benchmarks/python_func_ops_example/rank-1.Jan_17_19_32_44.5411.pt.trace.json.gz]: Mean +- std dev: 96.5 sec +- 1.6 sec
parse[ijson:../../load_benchmarks/python_func_ops_example/rank-1.Jan_17_19_32_44.5411.pt.trace.json.gz]: Mean +- std dev: 45.5 sec +- 0.4 sec
parse[ijson_batched:../../load_benchmarks/python_func_ops_example/rank-1.Jan_17_19_32_44.5411.pt.trace.json.gz]: Mean +- std dev: 47.1 sec +- 0.4 sec
parse[ijson_batch_and_compress:../../load_benchmarks/python_func_ops_example/rank-1.Jan_17_19_32_44.5411.pt.trace.json.gz]: Mean +- std dev: 36.8 sec +- 0.6 sec
(hta) bcoutinho@bcoutinho-mbp HolisticTraceAnalysis2 % z
```
Full output 
https://gist.github.com/briancoutinho/d179890c464161a1f9fe8e56f90aae9c
  

## Before submitting

- [x] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [ ] N/A
- [x] Did you write any new necessary tests?
  - [ ] N/A
- [ ] Did you make sure to update the docs?
  - [x] N/A
- [ ] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [ ] N/A