@fengxizhou 2024-03-23
Added a comment to consolidate the two new parser options into the ParserConfig argument to keep the method signature simple and easy to extend in the future.